Changed k from long int to int, cannot be a long int anyways

+++ classed_version/gamma.cpp
@@ -81,8 +81,7 @@ double GammaCalculator::integrated_bispec(double psi, double phi, double A3){
     double spsi = sin(psi);
     double r1,r2,temp;
 
-    // WHY IS k an unsigned long? prec_k is a regular int, so this does not make too much sense...
-    for(unsigned long int k=1;k<prec_k;k++){
+    for(int k=1;k<prec_k;k++){
         r1 = array_psi[k]/A3*cpsi;
         r2 = array_psi[k]/A3*spsi;
         temp = bispectrum(r1, r2, phi)*array_product[k];
